AI Agents/requirements.txt
google-generativeai
requests
PyPDF2
python-dotenv
flask
flask-cors
